Avoid returning an all brokers down error on planned disconnections (#5126)
* Skip increasing brokers down count
for planned disconnections, those that were logged as DEBUG, including idle disconnections.
Only the first time it reaches the DOWN state it's skipped but it's counted if there are further disconnections.

* Avoid returning an all brokers down error on planned disconnections
This is done by avoiding to count planned disconnections, such as idle  disconnections, broker host change and similar, as events that can cause the client to reach the "all brokers down" state, returning an error and possibly restarting a re-bootstrap sequence since 2.10.0

* Add more information to disconnection error messages

* Skip mock test when using SSL

* `rd_atomic32_set` and `rd_atomic64_set`
return the previous value to be able to check if the atomic set changed the value

* Fix for `rd_kafka_broker_set_error`

* Regarding broker errors, only log warnings and only report errors to the user for levels less or equal to `ERR`

* Reduce flakyness: test 0083

* Reset the `rkb_down_reported` field when
a broker reaches the `UP` state, to check again other brokers before reaching the `ALL_BROKERS_DOWN` state. Avoids repeated `ALL_BROKERS_DOWN`, maybe triggered by idle disconnections when only connecting to a single broker

* Consider the disconnection after timeout
as a planned disconnection that doesn't count for reaching the all brokers down state

* Reset any broker down reported on re-bootstrap too

* Revert `rd_kafka_connect_any` to the previous version. Given we reset the broker down reported state it's not needed anymore to prioritize the learned brokers as all brokers will be tried again before starting a new re-boostrap sequence

* Add the `rk_rebootstrap_in_progress` field to prevent duplicated re-bootstrap starts if the initial re-bootstrap sequence isn't completed yet.

* Remove flakyness from test 0151 - `do_test_down_then_up_no_rebootstrap_loop` and make it consistent with current changes

* Fix test 0034
given the down reported state is reset, after retrying on re-bootstrap a new ALL_BROKERS_DOWN error is issued.
This happens after 1 second max (`rd_kafka_broker_reconnect_backoff`) and the test 0034 timeout is 5 seconds so it doesn't timeout and set the broker up again.
With an explicit timeout check the test is reliable.

* Fix memory leaks in tests

* Fix flakyness in test 0075 when using valgrind due to timing and multiple connect requests

* Set up a fixed max jitter of 2s

* Remove flakyness from test 0086_purge_remote

* Fix for a deadlock when calling `rd_kafka_filter_broker_by_GetTelemetrySubscription` introduced with #5130 (not released)