[hdrhistogram]: Prevent an FPE in rd_hdr_histogram_new (#4949)
The Floating Point Exception (FPE) `FE_DIVBYZERO` is flagged in
`rd_hdr_histogram_new` when `log2()` is called with `0`. This is done in
`rd_avg_init` numerous times, some of which cannot be disabled in
recent versions.

Consumers of librdkakfa may enable the raising of exceptions when FPEs
occur, including `FE_DIVBYZERO`, which causes the function to fail.

It is reasonable for the callers of `rd_hdr_histogram_new` to pass `0`
for the `minValue` parameter, and the intention of the calculation of
`unitMagnitude` is also reasonable. As such, fixing the maths there is
the correct approach. `unitMagnitude` being `0` in this case is an apt
choice based on how log2(n) works for n near `0`, and that the
caluclation already minimises to `0`.

A simple test was included that resets the flags, runs the offending
function, and observes the new set flags. Appropriately, it fails
before the fix, and passes after.