FIX: Improve PRNG seeding on Windows to ensure uniqueness of generated numbers (#5265)
Fix for the pseudo-random seed on Windows. The function `rand_r` isn't present on Windows and the global seed wasn't based on the current microseconds and thread id. Also it wasn't called on every thread as required on this platform but only once per process. The fix allows on this platform the uniqueness of client side member id generation in next-generation consumer group protocol.
Happening since 1.x

* Multiple platforms secure random generation
* Run KIP-848 tests on MinGW-w64
* Changes to srand per thread on Windows.
It's necessary on Windows to avoid using the same jitter values for all clients and avoid a stampede effect on brokers.
* Run macOS local quick tests with KIP-848 as well
* Run KIP-848 local quick tests on arm64 glibc and alpine
* Increase 0153 with a different number of maximum open files